# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JobTrack API",
    "description": "\nJobTrack API\n\nA production-oriented job application tracking backend with:\n\n- owner-scoped Companies and Jobs\n- Applications with validated status transitions and history\n- timezone-aware Interview scheduling and overlap detection\n- Dashboard funnel metrics with resilient Redis caching\n- JWT authentication, rotating refresh tokens, RBAC, and optional MFA/OIDC\n- PostgreSQL transactions, Alembic migrations, and operational telemetry\n",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/v1/interviews": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "summary": "Create Interview",
        "operationId": "create_interview_api_v1_interviews_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "List Interviews",
        "operationId": "list_interviews_api_v1_interviews_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Application Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InterviewStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "scheduled_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheduled From"
            }
          },
          {
            "name": "scheduled_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scheduled To"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/upcoming": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "List Upcoming Interviews",
        "operationId": "list_upcoming_interviews_api_v1_interviews_upcoming_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponse"
                  },
                  "title": "Response List Upcoming Interviews Api V1 Interviews Upcoming Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interviews/{interview_id}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "summary": "Get Interview",
        "operationId": "get_interview_api_v1_interviews__interview_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Interviews"
        ],
        "summary": "Update Interview",
        "operationId": "update_interview_api_v1_interviews__interview_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Interview Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Interviews"
        ],
        "summary": "Delete Interview",
        "operationId": "delete_interview_api_v1_interviews__interview_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "saved",
          "applied",
          "screening",
          "interview",
          "offer",
          "rejected",
          "withdrawn",
          "archived"
        ],
        "title": "ApplicationStatus"
      },
...
      "ApplicationSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "job_summary": {
            "$ref": "#/components/schemas/JobSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "status",
          "priority",
          "job_summary"
        ],
        "title": "ApplicationSummary"
      },
...
      "CompanySummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CompanySummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewCreate": {
        "properties": {
          "application_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Application Id"
          },
          "interview_type": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "type": "integer",
            "maximum": 480.0,
            "minimum": 15.0,
            "title": "Duration Minutes",
            "default": 60
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "application_id",
          "interview_type",
          "scheduled_at"
        ],
        "title": "InterviewCreate"
      },
      "InterviewListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InterviewResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "InterviewListResponse"
      },
      "InterviewResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "interview_type": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "status": {
            "$ref": "#/components/schemas/InterviewStatus"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "type": "integer",
            "title": "Duration Minutes"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "application_summary": {
            "$ref": "#/components/schemas/ApplicationSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "interview_type",
          "status",
          "scheduled_at",
          "duration_minutes",
          "created_at",
          "updated_at",
          "application_summary"
        ],
        "title": "InterviewResponse"
      },
      "InterviewStatus": {
        "type": "string",
        "enum": [
          "scheduled",
          "completed",
          "cancelled"
        ],
        "title": "InterviewStatus"
      },
      "InterviewType": {
        "type": "string",
        "enum": [
          "phone_screen",
          "technical",
          "behavioral",
          "system_design",
          "hiring_manager",
          "onsite",
          "other"
        ],
        "title": "InterviewType"
      },
      "InterviewUpdate": {
        "properties": {
          "interview_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewType"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InterviewStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 480.0,
                "minimum": 15.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "meeting_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meeting Url"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "InterviewUpdate"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "open",
          "paused",
          "closed"
        ],
        "title": "JobStatus"
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "company_summary": {
            "$ref": "#/components/schemas/CompanySummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "title",
          "status",
          "company_summary"
        ],
        "title": "JobSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/login/"
          }
        }
      }
    }
  }
}